requirements.txt
instaloader
openpyxl
requests
Pillow==10.0.0
selenium==4.14.0
webdriver-manager==4.0.0